Use multiple workers instead of a single server
Simply copy the _server.lua once or twice in the
scripts directory. They will register themselves with
the "client" automatically. Props to anon for reminding me.

This improves performance a bunch, at the cost of,
well, peformance. Your CPU will max out.

It'd be cool to automatically generate the workers
from the main script, but you can't tell mpv to load
up a script, as far as I know.