Playlist logn EDL paths, or give up
Some ytdl Youtube EDLs can be even 150 kilobytes long, which is much
too much for commandline and even playlist files (mpv can handle 8k
characters on one line).

If a file path (url) is over 8k characters long, use original path
(url) and permit the use of ytdl. This will be slow, but it will work.
If a file path is over 1k characters long (but less than 8k), save it
to a temporary playlist and play that. 1k may be too long for
Windows commandline; I didn't really check.