Improve thumbnail feel with a half-offset
Previously, thumbnails were created precisely on the mark.
Thumbnail 1 (0) would be at 0 seconds, thumbnail 2 at 5 seconds,
3 at 10 etc.
This meant that when the user hovered over a thumbnail, it was already
a past event. This felt off, because the visible thumbnail was never
properly accurate.

Now, we take thumbnails half a delta later. Thumbnail 1 will be at 2.5
seconds, 2 at 7.5, 3 at 12.5. This means that when moving the mouse from
left to right, the thumbnail changes 2.5 seconds "before" the the
thumbnailed timestamp. The thumbnail is no longer guaranteed to be a past
event, and feels much better.

A random flick towards the seekbar means a 50% chance that the seek will
be before the displayed moment. We *could* always generate/display the
"next" moment (1 = 5s, 2 = 10s, etc), but this sort of offset feels
slightly wrong as well. For now, we'll make do with the 50% offset.

Fixes #7.